Disable hide notification in live viewer and set hide notification in live boardcaster instead
